server/requirements.txt
# Application requirements
django
django-debug-toolbar==1.9.1
django-environ==0.4.4
djangorestframework==3.8.2
django-radius==1.3.3
drf-yasg==1.6.1
gunicorn==19.7.1
psycopg2-binary==2.7.4
python-dotenv==0.8.2
gevent==1.2.2
django-filter==2.0.0
djangorestframework-jsonapi==2.7.0
django-extra-fields==1.0.0
python-jose==3.0.1
pycryptodome==3.7.0
pypdf2==1.26.0
requests==2.21.0
celery==4.3.0
redis==3.2.1
django-cors-headers==2.5.2
iso8601

# Development requirements
coverage==4.5.1
isort==4.3.4
pycodestyle==2.4.0
pylint-mccabe==0.1.3
pytest==3.5.0
pytest-html==1.17.0
pytest-django==3.2.1
faker==1.0.0
